move function to ci_visibility namespace
